requirements.txt
streamlit
openai
pillow
pytesseract
PyPDF2
python-docx
duckduckgo-search
pyyaml
requests
together
groq
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
pydantic
asyncio
aiohttp